[FAB] Improve the shortcut dialog ux after change accessibility button mode
Problem: The "Customize accessibility button" link only exists when the user selects the FAB location, which may confuse and disorient the user
Solution: It is hard to back to floating menu mode if we turns to navigation bar from "Customize accessibility button" setting.  Added the link in the "navigation bar mode" can help on this problem.

Fix: 183342563
Test: manual test
Change-Id: Ife572b0962f3a57c34af56cd930a487c9d813c0c

Former-commit-id: 54ded4cb5e717051b070eeaf177cb941e30e97a7