Add TextWatcher for the ip config field
[Root Cause]
We always check the valid for the ip config field. If the text is empty
we will set the default for user even if user deleted the values.

[Solution]
Add TextWatcher for the ip config field (include gateway/network prefix
length/dns1). If the text change and the text is empty, we will set the
hint for user, not the default.

Bug: 179331339
Test: manual test
       make RunSettingsRoboTests ROBOTEST_FILTER=WifiConfigController2
Change-Id: I1976200c70ae768285a44aff5df54ced00a78171

Former-commit-id: 244268b98d37d81f0bfe2ad470542e334201c7d8