Place GeoTZ toggle under auto time zone toggle.
This moves the two toggles that control the time zone detection
behavior next to each other, rather than having them either side
of the current device time zone. This is intended to making it
clear they are both influencing detection behavior. It will ensure
that future text we put under the auto time zone toggle for the
tablet use case (where there is no telephony detection) explaining
that "Use location to set time zone" has to be on will be right
next to the toggle that needs to change.

Bug: 152746236
Bug: 186625820
Test: flashed device, checked manually
Change-Id: Ifd43db3d5cc7e764b12f8da20e552b1a5479ac4a

Former-commit-id: ac584f2911674822114e49fba9f2dc74e01485e2