Fixes 'no ripple effect' issue for screen attention setting
Preferences shouldn't be initialized at onAttach() because the settings
style hasn't been loaded yet. This change defers the preferences
initialization so that they comply with the settings style.

The change ag/14114425 is reverted due to b/186904496. This change
includes a fix for that issue.

Test: manually
Test: make RunSettingsRoboTests
ROBOTEST_FILTER=com.android.settings.display
Bug: 183909540
Bug: 186904496

Change-Id: I317ce251f4d11e62c6592ee587c2919da4d45db3

Former-commit-id: 18c97b595fbb32de496103caa4e6379f69aa5e78