Fix searching condition for biometric settings
The preferences that are related to biometric settings should only be
searchable if the device supports the multiple biometric auth. However
we can see these preferences still can be searched even the device
doesn't support the feature, which is caused by that there's no
guardence. So the availiability of the preferences need to be updated to
prevent revealing unsupported preference from Settings search.

Fixes: 187167774
Test: manual test
1) Navigate to Settings search on a device without face auth
2) Search for face
3) See if there's something about face unlock

Change-Id: I6f120e8d9002a1e8f0be9361fdbea193776e4d7d

Former-commit-id: ecd149597709f207a863588ef98bc67b33cfdc75