Fix loading ring overlapping problem
By design, the loading ring is in the middle of the recycler view.
In the Android S, recyclerview isn't matched the screen's height.
So, if the recyclerview doesn't contains any item, the height could be
very low and then it causes to overlaps with drop down view.

The simple solution is to fix height of view of progress
bar, so it won't rely on the height of recycler view anymore.

Fix: 184643347
Test: Go to app notification page, and no overlapping problem.
Change-Id: I9f4fddba0254ac0cb0ba0b128a7dcc2de05d1fce

Former-commit-id: 2042dac123317d51ab449dab5baae7f830de5bb2