Add require scrolling button for FingerprintEnrolling intro
The consent in the fingerprint introduction has been updated. Users are
required to read through the content. Adding a scrolling button is to
prevent directly going to next page from this page. It also makes sure
the content has to be scrolled down to the bottom of page.

Fix: 184591438
Test: visual verified
1) Prepare a device with fingerprint sensor
2) Settings > Security > Fingerprint unlock
3) Entering the enrollment flow and see if the introduction has a button
named "More" on the right-bottom side.

Change-Id: I1916293807ca9fb04f5d576cf4f6aaccbbe1ff00

Former-commit-id: 872db9c41a285bab6b6d891bac4e20c33cecbf21