Apply transition on the back arrow on NotificationHistory page
When clicking on the back arrow from the action bar in the Notification
History page, the shared x-axis transition won't work.

The onOptionsItemSelected() overriden in the parent activity was
removed, so onNavigateUp() will get called. In the
NotificationHistoryActivity, onNavigateUp() has been overriden, and so
it will just call finish without any transition applied.

Hence, remove the override and just use the one define in
CollapsingToolbarBaseActivity where finishAfterTransition() is called.

Test: rebuild and click on the back arrow to see the transition
Fixes: 188126675
Change-Id: I34ff38ffcfb37d0471d325f2c31b726809fbd91f

Former-commit-id: 3e9c576bb8334f4bc1b25569eb913b7cf64045e2