Fix footer preference may flash one time when entering
new solution > We will not show the footer preference in the beginning when users entering the app, it will wait until the data is loaded to show the footer and app item in the same time. We also enhance the footer message as "Battery usage data will be available after you use your phone for a few hours" when there are not content in the begining.

Bug: 187766862
Bug: 188871770
Test: make SettingsRoboTests
Change-Id: I8d19173b4402722797b1ae3c4cfe6dcdf2d3fe3f

Former-commit-id: e030d8613a92ac6ade8a8ce3f62ef9ab704f7590