Fix a flicker problem on top app bar
The height of top app bar area is dependent on the line count of
multiline title. This mechanism is always triggered when entering either
subsetting page or rotating the screen, regardless the state of
collapsing toolbar. That is saying this mechanism caused a flicker
problem or a title in wrong color. The mechanism will be updated to
check the state of collapsing toolbar to make sure that it is only
triggered if the collapsing toolbar is in expanded state.

Fix: 186419089
Fix: 187709056
Test: visual verified
1) Navigate to a subsetting page with multiline title
2) Rotate the screen to landscape mode and collapse the title
3) Rotate the screen to portrait mode and see if the title is in
collapsed state

Change-Id: I3da2403fa43913469e64bd37e761b41a083efd38

Former-commit-id: fb2bfac434425b10323c3da300908b8e2a64de55