Align with the visibility logic from Dialer
Copy most visiblity logic from the Dialer app
We should check whether RTT is supported or not by querying Telecom.

Also, the rtt configuration won't be initialized properly before
user goes into the RTT page.
Prior to this cl, we set a wrong default value when we're showing
the subtext of RTT, this is not correct.
Currently, we simple show nothing in subtext if RTT haven't initilized
properly.

Test: Work with testing team to verify the behavior.
Fix: 188110773
Change-Id: I47c1983d15e22ef5a1fb3027334117a38a479558

Former-commit-id: 770f064d36025d48494c5207c2032aaf7664f72a