LegacyVpnPreferenceTest added for Vpn settings
Ensure that the profile name is successfully stored, and that the title
of the preference corresponds to the name of the associated VpnProfile

Interesting issue occurred where the icon resource was not present the
first time the class was initialized, but not the second. Trying this
twice within the test resolved this issue.

Test: atest -c LegacyVpnPreferenceTest
Bug: 187245804
Change-Id: I325b155cb0afdb6ce134325944fe5651bf9a1233

Former-commit-id: 74597578ca8578821a034fa888c4522cc99eb096