Add TARE button and homepage to developer options.
A TARE button is needed in the developer options in settings to allow
users and developers to change the default settings it will come with.
When clicked, it opens an activity of the TARE homepage with a toggle
and the two policies (Alarm Manager and Job Scheduler) to choose between. It also contains a revert to default settings button.

Bug: 191876567

Test: Manual (open Settings > System > Developer Options > TARE and
verify tabs and menus are there)

Change-Id: Id14bc9c1dcbaaf2aeb74ba7efb0168ab424f5d0c

Former-commit-id: 6ffa9dad04e525f6cda213620c6bf5edcf45a607