Provides TC ID for shortcut title in AccessibilityShortcutPreferenceFragment
Goal: Improve readability which is a bit difficult because of string concatenation for accurate translation.
Root cause: In some locales, framework features name would be a word rather than a product name. Hence it need to be in the different position in a sentence.
Solution: Use whole sentence to translate instead of concatenating the string could solve this issue.

Bug: 185478543
Test: Manually testing
Change-Id: Iced1fe7db81e4e78ddfed30ab947d449b5f9c735

Former-commit-id: 6c7b52b7030bde5a95ff39da61014ed4d5de4860