Create Dialog to edit TARE factors
This CL is the base framework for creating and adjusting TARE factor
values. TareFactorDialogFragment sets the basic outline for a factor
that uses a DialogBox to adjust the value. It takes the "title" or name
of the factor, the key that will be associated with the factor, and an
initial value for the factor; then sets up everything so that the factor
value can be easily changed and stored. These factors are also
registered into the settings observer class.

Bug: 195055619
Bug: 193238774

Test: Manual (Made a temporary class instance of the DialogFragment and
showed it in inactive apps onResume())

Change-Id: I3bb67717175e70eab3bbb11f4a42ee63a42f60bc

Former-commit-id: 5ebb245618f0887b7329887e903e7543534de855