Handle csip set member automatic pair in Setting
Move the set member handle to setting. SystemUI and Setting have the
individual settingLib instance. If we trigger pairing inside settingLib,
the action would be called twice. SystemUI does not contain the concept
to handle the pair interaction. Setting would take the action to
createBond for CSIP automatically.

Bug: 201758444
Bug: 150670922
Test: Pair the coordinated set devices, and check the set member pairing
would be done automatically without the pairing dialog

Change-Id: Idbdd21a1f15b04cc91cbc2d30a5cb4e9b24174b9

Former-commit-id: ebea73e36bdfd054416dfa868cfd70dfbdaaf79d