Receiving a broadcast message when home screen is ready.
There is a lag between Pixel/Phone is starting message is displayed and
when the home screen is actually ready. The lag is visible in low-end
devices like JioPhone Next. To solve that a broadcast message is sent
from the QuickstepLauncher class when the UI is ready and the
FallbackHome activity is finished listening to that broadcast message.

Bug: 199120420
Test: Tested on device

Change-Id: Ibc101a249c6dddcb890e8c3b75c9e27858844e9d

Former-commit-id: be02fd2ec3b388ba4946a4a3bb6d8eb71f1e43df