[Large screen] Shows MobileNetworkActivity in 2-pane
MobileNetworkActivity has launchMode singleTask, it makes
the Activity shows in full screen instead of 2-pane.

This change register SplitPairRule for the Activity to show
in 2-pane. However, there are other objects which starts
MobileNetworkActivity, it still needs later CL to refine
and register SplitPairRule for other objects which starts
MobileNetworkActivity.

Bug: 204039043
Test: manual
      Settings -> Network & Internet -> SIMs
Change-Id: Ia15c17c88efd46c5bb2b41aa21848293b9ff5610

Former-commit-id: 80887b6be3c043ed04f93a6c57fe6a1dfe2e0a4c